# openapi.yaml
openapi: 3.0.0
info:
  title: Blog API
  description: API documentation for a blog note-taking application.
  version: 1.0.0


paths:
  /api/blog:
    get:
      summary: Get all notes
      operationId: getNotes
      responses:
        '200':
          description: A list of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
    post:
      summary: Create a new note
      operationId: createNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewNote'
      responses:
        '201':
          description: Note created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'

  /api/blog/{id}:
    get:
      summary: Get a note by ID
      operationId: getNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Note retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Note not found
    patch:
      summary: Update a note by ID
      operationId: updateNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewNote'
      responses:
        '200':
          description: Note updated successfully
        '404':
          description: Note not found
        '500':
          description: Failed to update the note

    delete:
      summary: Delete a note by ID
      operationId: deleteNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Note deleted successfully
        '404':
          description: Note not found
        '500':
          description: Failed to delete the note

components:
  schemas:
    Note:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the note
        title:
          type: string
          description: The title of the note
        contents:
          type: string
          description: The contents of the note

    NewNote:
      type: object
      properties:
        title:
          type: string
          description: The title of the note
        contents:
          type: string
          description: The contents of the note
      required:
        - title
        - contents